requirements.txt
streamlit
pillow
pandas
matplotlib
seaborn
easyocr